:white_check_mark: Implement disposal of test logs and metrics
